requirements.txt
autopep8==1.5
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.42.0
django-crispy-forms==1.9.2
djangorestframework==3.11.0
idna==2.10
oauthlib==3.1.0
postgres==3.0.0
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pycodestyle==2.5.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.10
virtualenv==16.7.5
wincertstore==0.2
